pm: device: move internal function to internal header
Move internal function to internal header. This will ensure private
APIs are not accidentally used from drivers and the application.

To not break anything, update users of the internal functions to
include the internal header.
