+++ drivers/power_domain/power_domain_gpio.c
@@ -9,6 +9,7 @@
 
 #include <zephyr/kernel.h>
 #include <zephyr/drivers/gpio.h>
+#include <zephyr/internal/pm/device.h>
 #include <zephyr/pm/device.h>
 #include <zephyr/pm/device_runtime.h>
