+++ subsys/pm/device.c
@@ -5,7 +5,7 @@
  */
 
 #include <zephyr/device.h>
-#include <zephyr/pm/device.h>
+#include <zephyr/internal/pm/device.h>
 #include <zephyr/pm/device_runtime.h>
 #include <zephyr/sys/iterable_sections.h>
