+++ subsys/shell/modules/device_service.c
@@ -10,7 +10,7 @@
 #include <string.h>
 #include <stdio.h>
 #include <zephyr/device.h>
-#include <zephyr/pm/device.h>
+#include <zephyr/internal/pm/device.h>
 #include <zephyr/pm/device_runtime.h>
 #include <zephyr/arch/arch_interface.h>
